+++ src/TRegx/CleanRegex/Exception/NotReplacedException.php
@@ -3,8 +3,17 @@ namespace TRegx\CleanRegex\Exception;
 
 class NotReplacedException extends PatternException
 {
-    public function __construct(string $message)
+    /** @var string */
+    public $subject;
+
+    public function __construct(string $message, string $subject)
     {
         parent::__construct($message);
+        $this->subject = $subject;
+    }
+
+    public function getSubject(): string
+    {
+        return $this->subject;
     }
 }